+++ gui-builder/include/SpinButtonProperties.hpp
@@ -35,11 +35,11 @@ struct SpinButtonProperties : WidgetProperties
     {
         auto spinButton = std::dynamic_pointer_cast<tgui::SpinButton>(widget);
         if (property == "Minimum")
-            spinButton->setMinimum(tgui::stoi(value));
+            spinButton->setMinimum(tgui::stof(value));
         else if (property == "Maximum")
-            spinButton->setMaximum(tgui::stoi(value));
+            spinButton->setMaximum(tgui::stof(value));
         else if (property == "Value")
-            spinButton->setValue(tgui::stoi(value));
+            spinButton->setValue(tgui::stof(value));
         else
             WidgetProperties::updateProperty(widget, property, value);
     }
